test: Consolidated tests for mpsa.
